Create OpenWorkbook
The is a Macro which opens files that are usually saved in the same location by an application, but you don't know the username of the person trying to open it, in this case SAP GUI (SAP by default saves exported content to the SAP GUI folder of a computer if not otherwise specified). Then it tries different variants of usernames to open the file. This is useful if you make a Macro, and you want to open an exported file, but you don't know who the user will be who will be opening the file. The "Environ("Username") grabs the username of the individual from windows and places it in the cell X1 on the test page. Then it also adds a variant of the X1 username in X2 if needed (some orgs use multiple variants of usernames for users). 

This is definitely a lot of work as you could just ask the user to put their username in a cell in the workbook and then use a much shorter amount of code to open a file. But this does work pretty well if you want to share your macro with lots of people and not set it up for them.