Create Fast Delete Rows
If you need to delete rows in Excel based on a certain criteria, and you have 10K+ rows to do this with, other methods do not work. Or if they do they can take literal hours to complete. In order to quickly delete rows on large datasets, I use the FastDeleteRows Macro. The only thing to remember with this one is that it completes by creating a new sheet named "Paste1" which will contain the information that you want to keep. This is by far the fastest way to complete operations like this in Excel.